Fixed spacing, updated to preferred types.

+++ haskell-from-first-principles/test-files/FunctionWithLetAndWhere.hs
@@ -2,44 +2,44 @@
 module FunctionWithLetAndWhere where
 
 
-printInc :: Int -> IO()
+printInc :: Integer -> IO()
 printInc n = print plusTwo
   where plusTwo = n + 2
 
 
-printInc2 :: Int -> IO()
+printInc2 :: Integer -> IO()
 printInc2 n =
   let plusTwo = n + 2
   in print plusTwo
 
 
-rewrite1 :: Int
+rewrite1 :: Integer
 rewrite1  = x * 3 + y
   where x = 3
         y = 1000
 
 
-rewrite2 :: Int
+rewrite2 :: Integer
 rewrite2  = x * 5
   where y = 10
         x = 10 * 5 + y
 
 
-rewrite3 :: Float
+rewrite3 :: Double
 rewrite3  = z / x + y
   where x = 7
         y = negate x
         z = y * 10
 
 
-waxOn :: Int
+waxOn :: Integer
 waxOn     = x * 5
   where z = 7
         y = z + 8
         x = y ^ 2
 
 
-triple :: Int -> Int
+triple :: Integer -> Integer
 triple x = x * 3
